// src/app/shared/model/event.model.ts
import { ISession } from './event.session';

export interface IEvent {
    id: number;
    name: string;
    date: Date;
    time: string;
    price: number;
    imageUrl: string;
    location?: {
        address: string;
        city: string;
        country: string;
    };
    onlineUrl?: string;
    sessions: ISession[];
}